update `saveFigureSpeech()`
change output file names to work with stylo

+++ R/save.R
@@ -6,7 +6,6 @@
 #' @param min_token_count The minimal token count for a speech to be saved in a file (default = 0)
 #' @param count_punctuation Whether to include punctuation in min_token_count
 #' @param dir The directory into which the files will be written (default: current working directory)
-#' @param write_drama_name TODO!
 #' @export
 #' @examples
 #' \dontrun{
@@ -17,15 +16,32 @@ saveFigureSpeech <- function(t,
                              segment=c("Drama", "Act", "Scene"),
                              min_token_count=0,
                              count_punctuation=TRUE,
-                             dir=getwd(),
-                             write_drama_name=TRUE) {
+                             dir=getwd()) {
+  t <- t[, Token.surface, by=.(drama, Speaker.figure_id, Number.Act, Number.Scene)]
+  t$drama <- gsub("_", ".", t[,drama])
+  t$Speaker.figure_id <- gsub("_", ".", t[,Speaker.figure_id])
   segment = match.arg(segment)
   switch(segment,
-         Drama = {r <- split(t[, Token.surface, by=.(drama, Speaker.figure_id)], by=c("drama", "Speaker.figure_id"), keep.by = FALSE, drop=TRUE)},
-         Act = {r <- split(t[, Token.surface, by=.(drama, Speaker.figure_id, Number.Act)], by=c("drama", "Speaker.figure_id", "Number.Act"), keep.by=FALSE, drop=TRUE)},
-         Scene = {r <- split(t[, Token.surface, by=.(drama, Speaker.figure_id, Number.Act, Number.Scene)], by=c("drama", "Speaker.figure_id", "Number.Act", "Number.Scene"), keep.by=FALSE, drop=TRUE)},
+         Drama = {
+           t <- t[,fn:=paste(drama, Speaker.figure_id, sep="_")]
+         },
+         Act = {
+           if (length(unique(t$drama))==1) {
+             t <- t[,fn:=paste(Speaker.figure_id, Number.Act, sep="_")]
+           } else {
+             t <- t[,fn:=paste(drama, Speaker.figure_id, Number.Act, sep="_")]
+           }
+         },
+         Scene = {
+           if (length(unique(t$drama))==1) {
+             t <- t[,fn:=paste(Speaker.figure_id, Number.Act, Number.Scene, sep="_")]
+           } else {
+             t <- t[,fn:=paste(drama, Speaker.figure_id, Number.Act, Number.Scene, sep="_")]
+           }
+         },
          stop("Please enter valid string-value for argument 'segment' (default = 'Drama', 'Act' or 'Scene').")
   )
+  r <- split(t[, c("Speaker.figure_id", "Number.Act", "Number.Scene"):=NULL], by="fn", keep.by=FALSE, drop=TRUE)
   if (count_punctuation) {
     o <- capture.output(lapply(names(r), function(x) {
       if (length(r[[x]]$Token.surface) >= min_token_count){
@@ -35,10 +51,12 @@ saveFigureSpeech <- function(t,
       }
     }))
   } else {
-    o <- capture.output(lapply(names(r), function(x) {if (length(r[[x]][!grep(pattern="[[:punct:]]", x=r[[x]]$Token.surface)]$Token.surface) >= min_token_count) 
-      fn <- paste(dir, "/", x, sep="")
-      if (file.exists(fn)) {file.remove(fn)}
-      lapply(r[[x]], cat, sep="\n", file=fn, append=TRUE)
+    o <- capture.output(lapply(names(r), function(x) {
+      if (length(r[[x]][!grep(pattern="[[:punct:]]", x=r[[x]]$Token.surface)]$Token.surface) >= min_token_count){
+        fn <- paste(dir, "/", x, sep="")
+        if (file.exists(fn)) {file.remove(fn)}
+        lapply(r[[x]], cat, sep="\n", file=fn, append=TRUE)
+      }
     }))
   }
   message(length(grep(".*\\$Token\\.surface", x=o)), " files were written to ", dir)
